+++ components/number-keyboard/index.vue
@@ -144,4 +144,5 @@ export default {
     padding-top 1px
     background-color color-bg-base
     padding-bottom constant(safe-area-inset-bottom)
+    padding-bottom env(safe-area-inset-bottom)
 </style>